// Angular-2/Workshops/2. Angular/top-movies/top-movies_skeleton/app/core/models/movie.ts
export class MovieModel {
    Title: string;
    Year: number;
    Rated: string;
    Released: string;
    Runtime: string;
    Genre: string;
    Director: string;
    Writer: string;
    Actors: string;
    Plot: string;
    Language: string;
    Country: string;
    Awards: string;
    Poster: string;
    Metascore: string;
    imdbRating: number;
    imdbVotes: number;
    imdbID: string;
    Type: string;
    Top250: number;
}